tests/wpt/web-platform-tests/css/css-scroll-anchoring/parsing/overflow-anchor-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Anchoring: parsing overflow-anchor with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-anchoring/#propdef-overflow-anchor">
<meta name="assert" content="overflow-anchor supports the full grammar 'auto | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("overflow-anchor", "auto");
test_valid_value("overflow-anchor", "none");
</script>
</body>
</html>
